+++ src/phi_force.c
@@ -32,11 +32,13 @@
 #include "kernel.h"
 #include "field_grad_s.h"
 #include "field_s.h"
-#include "hydro_s.h"
+#include "hydro.h"
 #include "pth_s.h"
 #include "timer.h"
 #include "phi_force.h"
 #include "phi_force_colloid.h"
+#include "physics.h"
+
 
 static int phi_force_compute_fluxes(lees_edw_t * le, fe_t * fe, int nall,
 				    double * fxe,
@@ -57,6 +59,8 @@ static __host__ int phi_force_wallx(cs_t * cs, wall_t * wall, fe_t * fe, double 
 static int phi_force_fluid_phi_gradmu(lees_edw_t * le, pth_t * pth, fe_t * fe,
 				      field_t * phi,
 				      hydro_t * hydro);
+static int phi_force_fluid_phi_gradmu_ext(cs_t * cs, field_t * fphi,
+					  hydro_t * hydro);
 
 /*****************************************************************************
  *
@@ -103,6 +107,7 @@ __host__ int phi_force_calculation(cs_t * cs, lees_edw_t * le, wall_t * wall,
       break;
     case PTH_METHOD_GRADMU:
       phi_force_fluid_phi_gradmu(le, pth, fe, phi, hydro);
+      phi_force_fluid_phi_gradmu_ext(cs, phi, hydro);
       break;
     default:
       assert(0);
@@ -180,6 +185,59 @@ static int phi_force_fluid_phi_gradmu(lees_edw_t * le, pth_t * pth,
         force[Z] = -phi*0.5*(mup1 - mum1);
 
 	/* Store the force on lattice */
+
+	hydro_f_local_add(hydro, index0, force);
+
+	/* Next site */
+      }
+    }
+  }
+
+  return 0;
+}
+
+
+/*****************************************************************************
+ *
+ *  phi_force_fluid_phi_gradmu_ext
+ *
+ *  As for phi_force_fluid_phi_gradmu(), except this is a contribution\
+ *  from the external chemical potential gradient.
+ *
+ *****************************************************************************/
+
+static int phi_force_fluid_phi_gradmu_ext(cs_t * cs, field_t * fphi,
+					  hydro_t * hydro) {
+  int ic, jc, kc;
+  int index0;
+  int nlocal[3];
+  double phi;
+  double force[3];
+
+  double grad_mu[3];
+  physics_t * phys = NULL;
+
+  assert(cs);
+  assert(fphi);
+  assert(hydro);
+
+  cs_nlocal(cs, nlocal);
+
+  physics_ref(&phys);
+  physics_grad_mu(phys, grad_mu);
+
+  for (ic = 1; ic <= nlocal[X]; ic++) {
+    for (jc = 1; jc <= nlocal[Y]; jc++) {
+      for (kc = 1; kc <= nlocal[Z]; kc++) {
+
+	index0 = cs_index(cs, ic, jc, kc);
+	field_scalar(fphi, index0, &phi);
+
+        force[X] = -phi*grad_mu[X];
+        force[Y] = -phi*grad_mu[Y];
+        force[Z] = -phi*grad_mu[Z];
+
+	/* Accumulate the force on lattice */
 
 	hydro_f_local_add(hydro, index0, force);