+++ src/subgrid.c
@@ -6,12 +6,10 @@
  *
  *  See Nash et al. (2007).
  *
- *  $Id$
- *
  *  Edinburgh Soft Matter and Statistical Phyiscs Group and
  *  Edinburgh Parallel Computing Centre
  *
- *  (c) 2010-2017 The University of Edinburgh
+ *  (c) 2010-2020 The University of Edinburgh
  *
  *  Contributing authors:
  *  Kevin Stratford (kevin@epcc.ed.ac.uk)
@@ -44,7 +42,9 @@ static int subgrid_on_ = 0;  /* Subgrid particle flag */
  *
  *****************************************************************************/
 
-int subgrid_force_from_particles(colloids_info_t * cinfo, hydro_t * hydro) {
+
+int subgrid_force_from_particles(colloids_info_t * cinfo, hydro_t * hydro,
+				 wall_t * wall) {
 
   int ic, jc, kc;
   int i, j, k, i_min, i_max, j_min, j_max, k_min, k_max;
@@ -52,21 +52,25 @@ int subgrid_force_from_particles(colloids_info_t * cinfo, hydro_t * hydro) {
   int nlocal[3], offset[3];
   int ncell[3];
 
-  double r[3], r0[3], force[3], g[3];
+  double r[3], r0[3], force[3];
   double dr;
   colloid_t * p_colloid;
 
-  physics_t * phys = NULL;
-
   assert(cinfo);
   assert(hydro);
+  assert(wall);
+
+  if (subgrid_on_ == 0) return 0;
 
   cs_nlocal(cinfo->cs, nlocal);
   cs_nlocal_offset(cinfo->cs, offset);
   colloids_info_ncell(cinfo, ncell);
 
-  physics_ref(&phys);
-  physics_fgrav(phys, g);
+  /* Add any wall lubrication corrections before communication to
+   * find total external force on each particle */
+
+  subgrid_wall_lubrication(cinfo, wall);
+  colloid_sums_halo(cinfo, COLLOID_SUM_FORCE_EXT_ONLY);
 
   /* Loop through all cells (including the halo cells) */
 
@@ -76,33 +80,35 @@ int subgrid_force_from_particles(colloids_info_t * cinfo, hydro_t * hydro) {
 
 	colloids_info_cell_list_head(cinfo, ic, jc, kc, &p_colloid);
 
-	while (p_colloid != NULL) {
+	for ( ; p_colloid; p_colloid = p_colloid->next) {
 
-          /* Need to translate the colloid position to "local"
-           * coordinates, so that the correct range of lattice
-           * nodes is found */
+          if (p_colloid->s.type != COLLOID_TYPE_SUBGRID) continue;
 
-          r0[X] = p_colloid->s.r[X] - 1.0*offset[X];
-          r0[Y] = p_colloid->s.r[Y] - 1.0*offset[Y];
-          r0[Z] = p_colloid->s.r[Z] - 1.0*offset[Z];
+	  /* Need to translate the colloid position to "local"
+	   * coordinates, so that the correct range of lattice
+	   * nodes is found */
+
+	  r0[X] = p_colloid->s.r[X] - 1.0*offset[X];
+	  r0[Y] = p_colloid->s.r[Y] - 1.0*offset[Y];
+	  r0[Z] = p_colloid->s.r[Z] - 1.0*offset[Z];
 
 	  /* Work out which local lattice sites are involved
 	   * and loop around */
 
-          i_min = imax(1,         (int) floor(r0[X] - drange_));
-          i_max = imin(nlocal[X], (int) ceil (r0[X] + drange_));
-          j_min = imax(1,         (int) floor(r0[Y] - drange_));
-          j_max = imin(nlocal[Y], (int) ceil (r0[Y] + drange_));
-          k_min = imax(1,         (int) floor(r0[Z] - drange_));
-          k_max = imin(nlocal[Z], (int) ceil (r0[Z] + drange_));
+	  i_min = imax(1,         (int) floor(r0[X] - drange_));
+	  i_max = imin(nlocal[X], (int) ceil (r0[X] + drange_));
+	  j_min = imax(1,         (int) floor(r0[Y] - drange_));
+	  j_max = imin(nlocal[Y], (int) ceil (r0[Y] + drange_));
+	  k_min = imax(1,         (int) floor(r0[Z] - drange_));
+	  k_max = imin(nlocal[Z], (int) ceil (r0[Z] + drange_));
 
-          for (i = i_min; i <= i_max; i++) {
-            for (j = j_min; j <= j_max; j++) {
+	  for (i = i_min; i <= i_max; i++) {
+	    for (j = j_min; j <= j_max; j++) {
 	      for (k = k_min; k <= k_max; k++) {
 
 		index = cs_index(cinfo->cs, i, j, k);
 
-                /* Separation between r0 and the coordinate position of
+		/* Separation between r0 and the coordinate position of
 		 * this site */
 
 		r[X] = r0[X] - 1.0*i;
@@ -110,17 +116,16 @@ int subgrid_force_from_particles(colloids_info_t * cinfo, hydro_t * hydro) {
 		r[Z] = r0[Z] - 1.0*k;
 
 		dr = d_peskin(r[X])*d_peskin(r[Y])*d_peskin(r[Z]);
-
-		force[X] = g[X]*dr;
-		force[Y] = g[Y]*dr;
-		force[Z] = g[Z]*dr;
+            
+		force[X] = p_colloid->fex[X]*dr;
+		force[Y] = p_colloid->fex[Y]*dr;
+		force[Z] = p_colloid->fex[Z]*dr;
 		hydro_f_local_add(hydro, index, force);
 	      }
 	    }
 	  }
 
 	  /* Next colloid */
-	  p_colloid = p_colloid->next;
 	}
 
 	/* Next cell */
@@ -147,7 +152,6 @@ int subgrid_update(colloids_info_t * cinfo, hydro_t * hydro) {
   int ic, jc, kc;
   int ncell[3];
   double drag, reta;
-  double g[3];
   double eta;
   PI_DOUBLE(pi);
   colloid_t * p_colloid;
@@ -155,6 +159,8 @@ int subgrid_update(colloids_info_t * cinfo, hydro_t * hydro) {
 
   assert(cinfo);
   assert(hydro);
+
+  if (subgrid_on_ == 0) return 0;
 
   colloids_info_ncell(cinfo, ncell);
 
@@ -165,7 +171,6 @@ int subgrid_update(colloids_info_t * cinfo, hydro_t * hydro) {
 
   physics_ref(&phys);
   physics_eta_shear(phys, &eta);
-  physics_fgrav(phys, g);
   reta = 1.0/(6.0*pi*eta);
 
   for (ic = 0; ic <= ncell[X] + 1; ic++) {
@@ -174,18 +179,18 @@ int subgrid_update(colloids_info_t * cinfo, hydro_t * hydro) {
 
 	colloids_info_cell_list_head(cinfo, ic, jc, kc, &p_colloid);
 
-	while (p_colloid != NULL) {
+	for ( ; p_colloid; p_colloid = p_colloid->next) {
+
+          if (p_colloid->s.type != COLLOID_TYPE_SUBGRID) continue;
 
 	  drag = reta*(1.0/p_colloid->s.a0 - 1.0/p_colloid->s.ah);
 
 	  for (ia = 0; ia < 3; ia++) {
-	    p_colloid->s.v[ia] = p_colloid->fc0[ia] + drag*g[ia];
+	    p_colloid->s.v[ia] = p_colloid->fsub[ia] + drag*p_colloid->fex[ia];
 	    p_colloid->s.dr[ia] = p_colloid->s.v[ia];
 	  }
-
-	  p_colloid = p_colloid->next;
 	}
-
+	/* Next cell */
       }
     }
   }
@@ -230,11 +235,13 @@ static int subgrid_interpolation(colloids_info_t * cinfo, hydro_t * hydro) {
 
 	colloids_info_cell_list_head(cinfo, ic, jc, kc, &p_colloid);
 
-	while (p_colloid != NULL) {
-	  p_colloid->fc0[X] = 0.0;
-	  p_colloid->fc0[Y] = 0.0;
-	  p_colloid->fc0[Z] = 0.0;
-	  p_colloid = p_colloid->next;
+	for ( ; p_colloid; p_colloid = p_colloid->next) {
+
+          if (p_colloid->s.type != COLLOID_TYPE_SUBGRID) continue;
+
+	  p_colloid->fsub[X] = 0.0;
+	  p_colloid->fsub[Y] = 0.0;
+	  p_colloid->fsub[Z] = 0.0;
 	}
       }
     }
@@ -248,33 +255,35 @@ static int subgrid_interpolation(colloids_info_t * cinfo, hydro_t * hydro) {
 
 	colloids_info_cell_list_head(cinfo, ic, jc, kc, &p_colloid);
 
-	while (p_colloid != NULL) {
+	for ( ; p_colloid; p_colloid = p_colloid->next) {
 
-          /* Need to translate the colloid position to "local"
-           * coordinates, so that the correct range of lattice
-           * nodes is found */
+          if (p_colloid->s.type != COLLOID_TYPE_SUBGRID) continue;
 
-          r0[X] = p_colloid->s.r[X] - 1.0*offset[X];
-          r0[Y] = p_colloid->s.r[Y] - 1.0*offset[Y];
-          r0[Z] = p_colloid->s.r[Z] - 1.0*offset[Z];
+	  /* Need to translate the colloid position to "local"
+	   * coordinates, so that the correct range of lattice
+	   * nodes is found */
+
+	  r0[X] = p_colloid->s.r[X] - 1.0*offset[X];
+	  r0[Y] = p_colloid->s.r[Y] - 1.0*offset[Y];
+	  r0[Z] = p_colloid->s.r[Z] - 1.0*offset[Z];
 
 	  /* Work out which local lattice sites are involved
 	   * and loop around */
 
-          i_min = imax(1,         (int) floor(r0[X] - drange_));
-          i_max = imin(nlocal[X], (int) ceil (r0[X] + drange_));
-          j_min = imax(1,         (int) floor(r0[Y] - drange_));
-          j_max = imin(nlocal[Y], (int) ceil (r0[Y] + drange_));
-          k_min = imax(1,         (int) floor(r0[Z] - drange_));
-          k_max = imin(nlocal[Z], (int) ceil (r0[Z] + drange_));
+	  i_min = imax(1,         (int) floor(r0[X] - drange_));
+	  i_max = imin(nlocal[X], (int) ceil (r0[X] + drange_));
+	  j_min = imax(1,         (int) floor(r0[Y] - drange_));
+	  j_max = imin(nlocal[Y], (int) ceil (r0[Y] + drange_));
+	  k_min = imax(1,         (int) floor(r0[Z] - drange_));
+	  k_max = imin(nlocal[Z], (int) ceil (r0[Z] + drange_));
 
-          for (i = i_min; i <= i_max; i++) {
-            for (j = j_min; j <= j_max; j++) {
+	  for (i = i_min; i <= i_max; i++) {
+	    for (j = j_min; j <= j_max; j++) {
 	      for (k = k_min; k <= k_max; k++) {
 
 		index = cs_index(cinfo->cs, i, j, k);
 
-                /* Separation between r0 and the coordinate position of
+		/* Separation between r0 and the coordinate position of
 		 * this site */
 
 		r[X] = r0[X] - 1.0*i;
@@ -284,20 +293,48 @@ static int subgrid_interpolation(colloids_info_t * cinfo, hydro_t * hydro) {
 		dr = d_peskin(r[X])*d_peskin(r[Y])*d_peskin(r[Z]);
 		hydro_u(hydro, index, u);
 
-		p_colloid->fc0[X] += u[X]*dr;
-		p_colloid->fc0[Y] += u[Y]*dr;
-		p_colloid->fc0[Z] += u[Z]*dr;
+		p_colloid->fsub[X] += u[X]*dr;
+		p_colloid->fsub[Y] += u[Y]*dr;
+		p_colloid->fsub[Z] += u[Z]*dr;
 	      }
 	    }
 	  }
 
 	  /* Next colloid */
-	  p_colloid = p_colloid->next;
 	}
 
 	/* Next cell */
       }
     }
+  }
+
+  return 0;
+}
+
+/*****************************************************************************
+ *
+ *  subgrid_wall_lubrication
+ *
+ *  Accumulate lubrication corrections to the external force on each particle.
+ *
+ *****************************************************************************/
+
+int subgrid_wall_lubrication(colloids_info_t * cinfo, wall_t * wall) {
+
+  double drag[3];
+  colloid_t * pc = NULL;
+
+  assert(cinfo);
+  assert(wall);
+
+  colloids_info_local_head(cinfo, &pc);
+
+  for ( ; pc; pc = pc->next) {
+    if (pc->s.type != COLLOID_TYPE_SUBGRID) continue;
+    wall_lubr_sphere(wall, pc->s.ah, pc->s.r, drag);
+    pc->fex[X] += drag[X]*pc->s.v[X];
+    pc->fex[Y] += drag[Y]*pc->s.v[Y];
+    pc->fex[Z] += drag[Z]*pc->s.v[Z];
   }
 
   return 0;