+++ src/subgrid.h
@@ -2,24 +2,26 @@
  *
  *  subgrid.h
  *
- *  $Id: subgrid.h,v 1.2 2010-10-15 12:40:03 kevin Exp $
  *
  *  Edinburgh Soft Matter and Statistical Physics Group and
  *  Edinburgh Parallel Computing Centre
  *
  *  Kevin Stratford (kevin@epcc.ed.ac.uk)
- *  (c) 2010 The University of Edinburgh
+ *  (c) 2010-2020 The University of Edinburgh
  *
  *****************************************************************************/
 
-#ifndef SUBGRID_H
-#define SUBGRID_H
+#ifndef LUDWIG_SUBGRID_H
+#define LUDWIG_SUBGRID_H
 
 #include "colloids.h"
 #include "hydro.h"
+#include "wall.h"
 
 int subgrid_update(colloids_info_t * cinfo, hydro_t * hydro);
-int subgrid_force_from_particles(colloids_info_t * cinfo, hydro_t * hydro);
+int subgrid_force_from_particles(colloids_info_t * cinfo, hydro_t * hydro,
+				 wall_t * wall);
+int subgrid_wall_lubrication(colloids_info_t * cinfo, wall_t * wall);
 int subgrid_on_set(void);
 int subgrid_on(int * flag);