+++ tests/unit/tests.h
@@ -6,7 +6,7 @@
  *  Edinburgh Parallel Computing Centre
  *
  *  Kevin Stratford (kevin@epcc.ed.ac.uk)
- *  (c) 2010-2019 The University of Edinburgh
+ *  (c) 2010-2020 The University of Edinburgh
  *
  *****************************************************************************/
 
@@ -45,6 +45,8 @@ int test_field_grad_suite(void);
 int test_halo_suite(void);
 int test_hydro_suite(void);
 int test_io_suite(void);
+int test_io_options_suite(void);
+int test_io_options_rt_suite(void);
 int test_le_suite(void);
 int test_kernel_suite(void);
 int test_lubrication_suite(void);
@@ -65,5 +67,6 @@ int test_random_suite(void);
 int test_rt_suite(void);
 int test_timer_suite(void);
 int test_util_suite(void);
+int test_visc_arrhenius_suite(void);
 
 #endif